requirements.txt
discord.py
aiohttp
lxml
forex_python
pyyaml
aiofiles